refactor(logic): split feedstock-extraction gates into part file (Refs #2847)
full_ai_civilian_work_selection.dart exceeded the 1000 non-comment line
limit (1014) after the seller feedstock-tile acquisition residual was
added, failing repo.dart_file_non_comment_line_size in the Quality check.
Move the feedstock-extraction resource-id gates and Old World feedstock-tile
ownership predicates into a new part file (part of the same library, so the
public ai_api exports and shared private scope are preserved). Behavior is
unchanged.
